apps/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyYAML==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2

# AI Model Dependencies (Enhancement 1.1: Real AI Model Integration)
# Audio: Whisper for speech-to-text
openai-whisper>=20231117
torch>=2.1.0

# Image: Tesseract OCR for text extraction
pytesseract>=0.3.10
Pillow>=10.0.0

# Vision: OpenAI/Anthropic for image understanding (optional)
openai>=1.0.0
anthropic>=0.18.0

# RAG Dependencies (Enhancement 1.2: Advanced RAG)
# Vector database
chromadb>=0.4.0

# Embeddings and re-ranking
sentence-transformers>=2.2.2

# Hybrid search (BM25)
rank-bm25>=0.2.2

# PDF processing (for document indexing)
pypdf>=3.17.0

# Future dependencies for other enhancements
# faiss-cpu==1.7.4
# paddleocr==2.7.3
# faster-whisper==0.10.0
# transformers==4.36.0